fixing migrations

+++ database/migrations/2022_11_14_192728_create_rooms_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('room_code');
             $table->string('address');
             $table->string('room_size');
-            $table->integer('hospital_id')->unsigned();
+            $table->bigInteger('hospital_id')->unsigned();
             $table->foreign('hospital_id')->references('id')->on('hospitals')->onDelete('cascade');
             $table->timestamps();
         });
